move zkpass stuff to zkpass namespace
